Minor renames + reformat
